+++ frontend/src/components/UserSettings/Subscription.tsx
@@ -15,38 +15,20 @@ import {
   ModalOverlay,
   ModalContent,
   ModalHeader,
-  ModalFooter,
-  ModalBody,
   ModalCloseButton,
+  ModalBody,
   useDisclosure,
-  Select,
-  FormControl,
-  FormLabel,
 } from "@chakra-ui/react"
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query"
-import { useState } from "react"
+import { useQuery } from "@tanstack/react-query"
 
-import {
-  type ApiError,
-  type SubscriptionUpdate,
-  type UserSubscription,
-  type SubscriptionPlan,
-  UsersService,
-  MiscService,
-} from "../../client"
+import { type SubscriptionPlan, MiscService } from "../../client"
 import useAuth from "../../hooks/useAuth"
-import useCustomToast from "../../hooks/useCustomToast"
-import { handleError, formatTimestamp } from "../../utils"
+import { formatTimestamp } from "../../utils"
+import PickSubscription from "./PickSubscription"
 
 const Subscription = () => {
-  const queryClient = useQueryClient()
-  const showToast = useCustomToast()
   const { user: currentUser } = useAuth()
   const { isOpen, onOpen, onClose } = useDisclosure()
-  const [selectedPlan, setSelectedPlan] = useState<string>("standard")
-  const [selectedPeriod, setSelectedPeriod] = useState<"monthly" | "annual">(
-    "monthly",
-  )
 
   // Fetch available subscription plans
   const plansQuery = useQuery({
@@ -56,41 +38,6 @@ const Subscription = () => {
 
   // Current subscription from user data
   const currentSubscription = currentUser?.subscription
-
-  // Mutation for creating/updating subscription
-  const subscriptionMutation = useMutation({
-    mutationFn: async (data: SubscriptionUpdate) => {
-      if (currentSubscription) {
-        const result = await UsersService.putUserSubscription({
-          requestBody: data,
-        })
-        return result
-      } else {
-        const result = await UsersService.postUserSubscription({
-          requestBody: data,
-        })
-        // Extract the subscription from the NewSubscriptionResponse
-        return result.subscription as UserSubscription
-      }
-    },
-    onSuccess: () => {
-      showToast("Success!", "Subscription updated successfully.", "success")
-      onClose()
-    },
-    onError: (err: ApiError) => {
-      handleError(err, showToast)
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ["users", "me"] })
-    },
-  })
-
-  const handleUpgrade = () => {
-    subscriptionMutation.mutate({
-      plan_name: selectedPlan as "free" | "standard" | "professional",
-      period: selectedPeriod,
-    })
-  }
 
   const capitalizePlanName = (planName: string) => {
     return planName.charAt(0).toUpperCase() + planName.slice(1)
@@ -212,67 +159,19 @@ const Subscription = () => {
         </Box>
       </Container>
 
-      {/* Upgrade modal */}
-      <Modal isOpen={isOpen} onClose={onClose} isCentered>
+      {/* Subscription picker modal */}
+      <Modal isOpen={isOpen} onClose={onClose} isCentered size="lg">
         <ModalOverlay />
         <ModalContent>
           <ModalHeader>Change subscription</ModalHeader>
           <ModalCloseButton />
-          <ModalBody>
-            <Stack spacing={4}>
-              <FormControl>
-                <FormLabel>Select plan</FormLabel>
-                <Select
-                  value={selectedPlan}
-                  onChange={(e) => setSelectedPlan(e.target.value)}
-                >
-                  {plans
-                    .filter((plan: SubscriptionPlan) => plan.name !== "free")
-                    .map((plan: SubscriptionPlan) => (
-                      <option key={plan.id} value={plan.name}>
-                        {capitalizePlanName(plan.name)}: ${plan.price}/month
-                      </option>
-                    ))}
-                </Select>
-              </FormControl>
-
-              <FormControl>
-                <FormLabel>Billing period</FormLabel>
-                <Select
-                  value={selectedPeriod}
-                  onChange={(e) =>
-                    setSelectedPeriod(e.target.value as "monthly" | "annual")
-                  }
-                >
-                  <option value="monthly">Monthly</option>
-                  <option value="annual">Annual (save TODO%)</option>
-                </Select>
-              </FormControl>
-
-              <Alert status="info">
-                <AlertIcon />
-                <Box>
-                  <AlertTitle>Billing information</AlertTitle>
-                  <AlertDescription>
-                    You will be redirected to our secure payment processor to
-                    complete your subscription.
-                  </AlertDescription>
-                </Box>
-              </Alert>
-            </Stack>
+          <ModalBody pb={6}>
+            <PickSubscription
+              user={currentUser}
+              onSuccess={onClose}
+              showHeading={false}
+            />
           </ModalBody>
-          <ModalFooter>
-            <Button variant="ghost" mr={3} onClick={onClose}>
-              Cancel
-            </Button>
-            <Button
-              variant={"primary"}
-              onClick={handleUpgrade}
-              isLoading={subscriptionMutation.isPending}
-            >
-              Proceed to payment
-            </Button>
-          </ModalFooter>
         </ModalContent>
       </Modal>
     </>
